refactor: 데이터 베이스 연결정보 패키지 경로 변경
