+++ src/main/java/chess/DbConnection.java
@@ -1,4 +1,4 @@
-package chess.domain.userAccess;
+package chess;
 
 import java.sql.Connection;
 import java.sql.DriverManager;
@@ -9,8 +9,8 @@ public class DbConnection {
     private static final String SERVER = "localhost:13306"; // MySQL 서버 주소
     private static final String DATABASE = "chess"; // MySQL DATABASE 이름
     private static final String OPTION = "?useSSL=false&allowPublicKeyRetrieval=true&serverTimezone=UTC";
-    private static final String USERNAME = "root"; //  MySQL 서버 아이디
-    private static final String PASSWORD = "root"; // MySQL 서버 비밀번호
+    private static final String USERNAME = "user"; //  MySQL 서버 아이디
+    private static final String PASSWORD = "password"; // MySQL 서버 비밀번호
 
     private DbConnection() {
     }
