+++ src/main/java/chess/domain/userAccess/room/RoomDao.java
@@ -9,7 +9,7 @@ import java.sql.SQLException;
 import java.util.ArrayList;
 import java.util.List;
 
-import static chess.domain.userAccess.DbConnection.getConnection;
+import static chess.DbConnection.getConnection;
 
 public class RoomDao {
